fix: float coords instead of int

+++ datatypes/Point.go
@@ -1,7 +1,7 @@
 package datatypes
 
 type Point struct {
-	M_X int
-	M_Y int
-	M_Z int
+	M_X float64
+	M_Y float64
+	M_Z float64
 }
